backend/requirements.txt
# Core Django
Django==6.0.1
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Authentication
djangorestframework-simplejwt==5.3.1

# AI Services
openai==1.55.0
google-generativeai
sentence-transformers
scikit-learn

# Async Processing
celery==5.4.0
redis==5.2.0

# Database
psycopg[binary]>=3.2.0
pgvector>=0.2.0

# Environment Variables
python-dotenv==1.0.1

# Server
gunicorn==23.0.0

# Image Processing (optional)
Pillow==11.0.0

# PDF Generation
reportlab==4.2.5

# API Documentation
drf-spectacular==0.28.0

# Monitoring
sentry-sdk[django]==1.38.0